Refactoring and and fixed custom query for mainview
Refactored the DataAccessObject to DataAccessObject instead of "ServiceDataBase"

Implemented a custom HQL query to get the list of customers to call for the mainview to avoid excessive use of bandwith between the database and the webserver if they were to be on different servers in the future. Before the whole list of customers were fetched and handled by the backend.  Only customers with a calldate today or earlier should appear.