Rewrote a lot of code and started implementing spring security
There was a lot of problems with the old codebase, the main issue was i was not using JpaRepositories in the old codebase whichh made the code harder to read and there was much more lines of code.

I have now changed to Spring boots autoconfiguration with hibernate.

I added a shoppingcart to each customer.

I started implementing Spring security. Users can now login (not register). Passwords are hashed in the database.

Next is to structure the application to handle more users withh a unique set of cusstomers and products. Also making sure a user does not have access to other users data.